Split up stepper indirection (#15111)
