+++ Marlin/src/HAL/HAL_STM32_F4_F7/STM32F7/TMC2660.cpp
@@ -37,7 +37,7 @@
 
 #include "../../../inc/MarlinConfig.h"
 #include "../../../Marlin.h"
-#include "../../../module/stepper_indirection.h"
+#include "../../../module/stepper/indirection.h"
 #include "../../../module/printcounter.h"
 #include "../../../libs/duration_t.h"
 #include "../../../libs/hex_print_routines.h"
