+++ Marlin/src/gcode/feature/trinamic/M906.cpp
@@ -26,7 +26,7 @@
 
 #include "../../gcode.h"
 #include "../../../feature/tmc_util.h"
-#include "../../../module/stepper_indirection.h"
+#include "../../../module/stepper/indirection.h"
 
 /**
  * M906: Set motor current in milliamps.
